requirements.txt
reportlab
pdfkit
streamlit
jinja2
pymysql
load_dotenv